Binding React components with backend data in the pages : HomePage + Navbar + ProductDetails
i had some issues in the productDetails styling but overall not bad

+++ database/migrations/2024_05_11_035154_create_reviews_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
             $table->foreignId('user_id')->constrained();
             $table->foreignId('product_id')->constrained();
             $table->text('body');
-            $table->smallInteger('rating');
+            $table->float('rating');
             $table->timestamps();
         });
     }
